move from archived to "poolType"
There are 3 potential "states" for a pool.

uni
sushi
archived

If a poolType property is missing in the pool declaration it treats it as a uni pool.

When this is set it reorganizes the pools into different tabs for each type. If a pool is marked as archived you will no longer be able to deposit to it via the UI.